+++ languagetool-standalone/CHANGES.md
@@ -5,6 +5,7 @@
 ...
 
 ### General
+  * The `warmup` setting for the config file, which had no effect anymore, has been removed
   * The deprecated `--word2vecmodel` and `--neuralnetworkmodel` options has been removed,
     as these features were not maintained and had never been used on languagetool.org